// frontend/lib/api/generated/model/comparativoConsumoOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */

export interface ComparativoConsumoOut {
  producto_id: number;
  producto_codigo: string;
  producto_nombre: string;
  fecha_inicio: string;
  fecha_fin: string;
  cantidad_teorica: number;
  cantidad_comprada: number;
  cantidad_recibida: number;
  cantidad_despachada: number;
}